Start flask with an sqlite3 database
Following this tutorial:
https://flask.palletsprojects.com/en/2.2.x/tutorial/factory/
Flask is now started through a create_app function, and is ready to
initialise its own sqlite3 database, rather than requiring a local
postgresql database set up by each user separately to this project.

There is not a database connected yet, but the code has been modified
ready so that one can be.