Move code into a package.
`static`, `templates`, and the python files are now inside a package `flaskr`.
This will allow us to make a test folder to test the code.

The paths in `mainApp` have been updated to make this work.

`__init__.py` is needed to make `flaskr` a package, but it can have nothing
in it. As long as we still run the program through `mainApp.py`, this is
fine. `flaskr` has to be a package, as otherwise flask can't recognise the
`templates` folder properly.

The .gitignore has also been updated to account for a location change of
the `__pycache__` files.