Split add_menu_item method into two methods.
Add a MenuItemModel class to store the data from the HTML forms in
objects.

Split the add_menu_item method in app.py into two methods - one that
receives the data from the form, checks that the data entered is valid,
and then creates a MenuItemModel object.

Then pass this object onto the next method, add_item, to add the item to
the database. This should allow for a testable method because we can
provide mock data, rather than requiring submission through the HTML
forms.