Change encryption type to Python's hash function.
The previous encryption type produced an array of bytes, which couldn't
be manually entered into the database to insert accounts to test with.

This has now been changed to use python's built-in hash function, so
that in the future, test accounts can be inserted into the database
along with the test menu data.