Refactor duplicated sql statements into a method.
The sql statements in waiter_order_confirm and waiter_order_delivered
are identical aside from different states. The sql statement is now in a
method where the state is passes as a parameter.

The formatting as a dictionary is also identical between the two
methods, so this is now also in it's own method, and called by both.